llm-app/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
faiss-cpu
PyPDF2
sentence-transformers
numpy